new task script added

+++ 0x02-functions_nested_loops/101-natural.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i;
-	int sum = 0;
+	long int sum = 0;
 
 	for (i = 1; i < 1024; ++i)
 	{
